--- MasterSuite.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd">
-<suite name="MasterSuite">
-  <test thread-count="5" name="MasterTest">
-    <classes>
-      <class name="com.LoginTest.VerifyLoginTestWithValidCredentials"/>
-    </classes>
-  </test> <!-- MasterTest -->
-</suite> <!-- MasterSuite -->
