+++ MasterTestSuite.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd">
+<suite name="MasterTestSuite">
+  <test thread-count="5" name="MasterTestSuite">
+    <classes>
+      <class name="com.LoginTest.VerifyLogin_Nil"/>
+    </classes>
+  </test> <!-- MasterTestSuite -->
+</suite> <!-- MasterTestSuite -->
