+++ test-output/MasterSuite/MasterTest.xml
@@ -1,12 +1,42 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!-- Generated by org.testng.reporters.JUnitXMLReporter -->
-<testsuite ignored="0" hostname="DESKTOP-QVM6J4J" failures="1" tests="1" name="MasterTest" time="5.92" errors="0" timestamp="2023-10-11T08:52:04 IST">
-  <testcase classname="com.LoginTest.VerifyLoginTestWithValidCredentials" name="TC_LoginTest_1_VerifyLoginTestWithValidCrdentials" time="1.011">
-    <failure type="java.lang.AssertionError" message="expected [FaceBook Homepage] but found [Log in to Facebook]">
-      <![CDATA[java.lang.AssertionError: expected [FaceBook Homepage] but found [Log in to Facebook]
-at com.LoginTest.VerifyLoginTestWithValidCredentials.TC_LoginTest_1_VerifyLoginTestWithValidCrdentials(VerifyLoginTestWithValidCredentials.java:27)
-at java.base/java.util.ArrayList.forEach(ArrayList.java:1540)
-... Removed 33 stack frames]]>
+<testsuite ignored="0" hostname="LAPTOP-J2KDJTSV" failures="0" tests="1" name="MasterTest" time="6.974" errors="0" timestamp="2023-10-12T10:13:17 IST">
+  <testcase classname="com.Utility.BaseClass" name="@BeforeMethod setUp" time="6.904">
+    <failure type="org.openqa.selenium.WebDriverException" message="target frame detached
+  (failed to check if window was closed: disconnected: Unable to receive message from renderer)
+  (Session info: chrome=117.0.5938.150)
+Build info: version: &amp;apos;4.10.0&amp;apos;, revision: &amp;apos;c14d967899&amp;apos;
+System info: os.name: &amp;apos;Windows 11&amp;apos;, os.arch: &amp;apos;amd64&amp;apos;, os.version: &amp;apos;10.0&amp;apos;, java.version: &amp;apos;17.0.7&amp;apos;
+Driver info: org.openqa.selenium.chrome.ChromeDriver
+Command: [d2a7d29794efc2669cd390a402efa93e, get {url=https://demo.guru99.com/V4/}]
+Capabilities {acceptInsecureCerts: false, browserName: chrome, browserVersion: 117.0.5938.150, chrome: {chromedriverVersion: 117.0.5938.149 (e3344ddefa1..., userDataDir: C:\Users\NILESH~1\AppData\L...}, fedcm:accounts: true, goog:chromeOptions: {debuggerAddress: localhost:61966}, networkConnectionEnabled: false, pageLoadStrategy: normal, platformName: windows, proxy: Proxy(), se:cdp: ws://localhost:61966/devtoo..., se:cdpVersion: 117.0.5938.150, setWindowRect: true, strictFileInteractability: false, timeouts: {implicit: 0, pageLoad: 300000, script: 30000}, unhandledPromptBehavior: dismiss and notify, webauthn:extension:credBlob: true, webauthn:extension:largeBlob: true, webauthn:extension:minPinLength: true, webauthn:extension:prf: true, webauthn:virtualAuthenticators: true}
+Session ID: d2a7d29794efc2669cd390a402efa93e">
+      <![CDATA[org.openqa.selenium.WebDriverException: target frame detached
+  (failed to check if window was closed: disconnected: Unable to receive message from renderer)
+  (Session info: chrome=117.0.5938.150)
+Build info: version: '4.10.0', revision: 'c14d967899'
+System info: os.name: 'Windows 11', os.arch: 'amd64', os.version: '10.0', java.version: '17.0.7'
+Driver info: org.openqa.selenium.chrome.ChromeDriver
+Command: [d2a7d29794efc2669cd390a402efa93e, get {url=https://demo.guru99.com/V4/}]
+Capabilities {acceptInsecureCerts: false, browserName: chrome, browserVersion: 117.0.5938.150, chrome: {chromedriverVersion: 117.0.5938.149 (e3344ddefa1..., userDataDir: C:\Users\NILESH~1\AppData\L...}, fedcm:accounts: true, goog:chromeOptions: {debuggerAddress: localhost:61966}, networkConnectionEnabled: false, pageLoadStrategy: normal, platformName: windows, proxy: Proxy(), se:cdp: ws://localhost:61966/devtoo..., se:cdpVersion: 117.0.5938.150, setWindowRect: true, strictFileInteractability: false, timeouts: {implicit: 0, pageLoad: 300000, script: 30000}, unhandledPromptBehavior: dismiss and notify, webauthn:extension:credBlob: true, webauthn:extension:largeBlob: true, webauthn:extension:minPinLength: true, webauthn:extension:prf: true, webauthn:virtualAuthenticators: true}
+Session ID: d2a7d29794efc2669cd390a402efa93e
+at org.openqa.selenium.remote.codec.w3c.W3CHttpResponseCodec.createException(W3CHttpResponseCodec.java:199)
+at org.openqa.selenium.remote.codec.w3c.W3CHttpResponseCodec.decode(W3CHttpResponseCodec.java:132)
+at org.openqa.selenium.remote.codec.w3c.W3CHttpResponseCodec.decode(W3CHttpResponseCodec.java:51)
+at org.openqa.selenium.remote.HttpCommandExecutor.execute(HttpCommandExecutor.java:191)
+at org.openqa.selenium.remote.service.DriverCommandExecutor.invokeExecute(DriverCommandExecutor.java:196)
+at org.openqa.selenium.remote.service.DriverCommandExecutor.execute(DriverCommandExecutor.java:171)
+at org.openqa.selenium.remote.RemoteWebDriver.execute(RemoteWebDriver.java:531)
+at org.openqa.selenium.remote.RemoteWebDriver.get(RemoteWebDriver.java:295)
+at com.Utility.BaseClass.setUp(BaseClass.java:27)
+at java.base/java.util.ArrayList.forEach(ArrayList.java:1511)
+... Removed 36 stack frames]]>
     </failure>
+  </testcase> <!-- @BeforeMethod setUp -->
+  <testcase classname="com.Utility.BaseClass" name="@AfterMethod tearDown" time="0.0">
+    <skipped/>
+  </testcase> <!-- @AfterMethod tearDown -->
+  <testcase classname="com.LoginTest.VerifyLoginTestWithValidCredentials" name="TC_LoginTest_1_VerifyLoginTestWithValidCrdentials" time="0.0">
+    <skipped/>
   </testcase> <!-- TC_LoginTest_1_VerifyLoginTestWithValidCrdentials -->
 </testsuite> <!-- MasterTest -->
