+++ test-output/MasterSuite/testng-failed.xml
@@ -8,6 +8,7 @@
           <include name="TC_LoginTest_1_VerifyLoginTestWithValidCrdentials"/>
           <include name="setUp"/>
           <include name="tearDown"/>
+          <include name="ObjectInit"/>
         </methods>
       </class> <!-- com.LoginTest.VerifyLoginTestWithValidCredentials -->
     </classes>
