+++ test-output/MasterTest/MasterTest.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!-- Generated by org.testng.reporters.JUnitXMLReporter -->
+<testsuite ignored="0" hostname="LAPTOP-J2KDJTSV" failures="0" tests="1" name="MasterTest" time="13.564" errors="0" timestamp="2023-10-12T10:18:35 IST">
+  <testcase classname="com.LoginTest.VerifyLoginTestWithValidData" name="TC_LoginTest_1_VerifyLoginTestWithValidCrdentials" time="3.203"/>
+</testsuite> <!-- MasterTest -->
