+++ test-output/MasterTestSuite/MasterTestSuite.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!-- Generated by org.testng.reporters.JUnitXMLReporter -->
+<testsuite ignored="0" hostname="LAPTOP-J2KDJTSV" failures="0" tests="1" name="MasterTestSuite" time="53.082" errors="0" timestamp="2023-10-12T11:57:23 IST">
+  <testcase classname="com.LoginTest.VerifyLogin_Nil" name="TC_LoginTest_001_For_Guru99" time="3.175"/>
+</testsuite> <!-- MasterTestSuite -->
