+++ test-output/junitreports/TEST-com.LoginTest.VerifyLoginTestWithValidData.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!-- Generated by org.testng.reporters.JUnitReportReporter -->
+<testsuite hostname="LAPTOP-J2KDJTSV" failures="0" tests="1" name="com.LoginTest.VerifyLoginTestWithValidData" time="13.518" errors="0" timestamp="2023-10-12T10:18:35 IST" skipped="0">
+  <testcase classname="com.LoginTest.VerifyLoginTestWithValidData" name="TC_LoginTest_1_VerifyLoginTestWithValidCrdentials" time="13.518"/>
+</testsuite> <!-- com.LoginTest.VerifyLoginTestWithValidData -->
