+++ test-output/junitreports/TEST-com.LoginTest.VerifyLogin_Nil.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!-- Generated by org.testng.reporters.JUnitReportReporter -->
+<testsuite hostname="LAPTOP-J2KDJTSV" failures="0" tests="1" name="com.LoginTest.VerifyLogin_Nil" time="53.036" errors="0" timestamp="2023-10-12T11:57:23 IST" skipped="0">
+  <testcase classname="com.LoginTest.VerifyLogin_Nil" name="TC_LoginTest_001_For_Guru99" time="53.036"/>
+</testsuite> <!-- com.LoginTest.VerifyLogin_Nil -->
