+++ test-output/old/MasterSuite/methods.html
@@ -1,10 +1,8 @@
 <h2>Methods run, sorted chronologically</h2><h3>&gt;&gt; means before, &lt;&lt; means after</h3><p/><br/><em>MasterSuite</em><p/><small><i>(Hover the method name to see the test class name)</i></small><p/>
 <table border="1">
 <tr><th>Time</th><th>Delta (ms)</th><th>Suite<br>configuration</th><th>Test<br>configuration</th><th>Class<br>configuration</th><th>Groups<br>configuration</th><th>Method<br>configuration</th><th>Test<br>method</th><th>Thread</th><th>Instances</th></tr>
-<tr bgcolor="8ff871">  <td>23/10/11 08:52:03</td>   <td>0</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&lt;&lt;BaseClass.tearDown()[pri:0, instance:com.LoginTest.VerifyLoginTestWithValidCredentials@1af2d44a]">&lt;&lt;tearDown</td> 
-<td>&nbsp;</td>  <td>main@20224131</td>   <td></td> </tr>
-<tr bgcolor="8ff871">  <td>23/10/11 08:51:58</td>   <td>-5115</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;BaseClass.setUp()[pri:0, instance:com.LoginTest.VerifyLoginTestWithValidCredentials@1af2d44a]">&gt;&gt;setUp</td> 
-<td>&nbsp;</td>  <td>main@20224131</td>   <td></td> </tr>
-<tr bgcolor="7e8a9e">  <td>23/10/11 08:52:02</td>   <td>-1019</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="VerifyLoginTestWithValidCredentials.TC_LoginTest_1_VerifyLoginTestWithValidCrdentials()[pri:0, instance:com.LoginTest.VerifyLoginTestWithValidCredentials@1af2d44a]">TC_LoginTest_1_VerifyLoginTestWithValidCrdentials</td> 
-  <td>main@20224131</td>   <td></td> </tr>
+<tr bgcolor="87b08d">  <td>23/10/12 10:13:10</td>   <td>0</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;BaseClass.setUp()[pri:0, instance:com.LoginTest.VerifyLoginTestWithValidCredentials@4116aac9]">&gt;&gt;setUp</td> 
+<td>&nbsp;</td>  <td>main@989447607</td>   <td></td> </tr>
+<tr bgcolor="87b08d">  <td>23/10/12 10:13:09</td>   <td>-924</td> <td title="&gt;&gt;BaseClass.ObjectInit()[pri:0, instance:com.LoginTest.VerifyLoginTestWithValidCredentials@4116aac9]">&gt;&gt;ObjectInit</td> 
+<td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td>  <td>main@989447607</td>   <td></td> </tr>
 </table>
