+++ test-output/old/MasterTest/MasterTest.properties
@@ -1,0 +1,1 @@
+[SuiteResult context=MasterTest]
