+++ test-output/old/MasterTest/index.html
@@ -1,0 +1,6 @@
+<html><head><title>Results for MasterTest</title></head>
+<frameset cols="26%,74%">
+<frame src="toc.html" name="navFrame">
+<frame src="main.html" name="mainFrame">
+</frameset>
+</html>
