+++ test-output/old/MasterTest/main.html
@@ -1,0 +1,2 @@
+<html><head><title>Results for MasterTest</title></head>
+<body>Select a result on the left-hand pane.</body></html>
