+++ test-output/old/MasterTest/toc.html
@@ -1,11 +1,11 @@
 <html>
 <head>
-<title>Results for MasterSuite</title>
+<title>Results for MasterTest</title>
 <link href="../testng.css" rel="stylesheet" type="text/css" />
 <link href="../my-testng.css" rel="stylesheet" type="text/css" />
 </head>
 <body>
-<h3><p align="center">Results for<br/><em>MasterSuite</em></p></h3>
+<h3><p align="center">Results for<br/><em>MasterTest</em></p></h3>
 <table border='1' width='100%'>
 <tr valign='top'>
 <td>1 test</td>
@@ -20,9 +20,9 @@
 <td><a target='mainFrame' href='reporter-output.html'>reporter output</a></td>
 <td><a target='mainFrame' href='testng.xml.html'>testng.xml</a></td>
 </tr></table>
-<table width='100%' class='test-skipped'>
+<table width='100%' class='test-passed'>
 <tr><td>
-<table style='width: 100%'><tr><td valign='top'>MasterTest (0/0/1)</td><td valign='top' align='right'>
+<table style='width: 100%'><tr><td valign='top'>MasterTest (1/0/0)</td><td valign='top' align='right'>
   <a href='MasterTest.html' target='mainFrame'>Results</a>
 </td></tr></table>
 </td></tr><p/>
