+++ test-output/old/MasterTestSuite/MasterTestSuite.properties
@@ -1,0 +1,1 @@
+[SuiteResult context=MasterTestSuite]
