+++ test-output/old/MasterTestSuite/groups.html
@@ -1,0 +1,1 @@
+<h2>Groups used for this test run</h2>
