+++ test-output/old/MasterTestSuite/main.html
@@ -1,0 +1,2 @@
+<html><head><title>Results for MasterTestSuite</title></head>
+<body>Select a result on the left-hand pane.</body></html>
