+++ test-output/old/MasterTestSuite/methods-not-run.html
@@ -1,0 +1,2 @@
+<h2>Methods that were not run</h2><table>
+</table>
