+++ test-output/old/MasterTestSuite/methods.html
@@ -1,0 +1,12 @@
+<h2>Methods run, sorted chronologically</h2><h3>&gt;&gt; means before, &lt;&lt; means after</h3><p/><br/><em>MasterTestSuite</em><p/><small><i>(Hover the method name to see the test class name)</i></small><p/>
+<table border="1">
+<tr><th>Time</th><th>Delta (ms)</th><th>Suite<br>configuration</th><th>Test<br>configuration</th><th>Class<br>configuration</th><th>Groups<br>configuration</th><th>Method<br>configuration</th><th>Test<br>method</th><th>Thread</th><th>Instances</th></tr>
+<tr bgcolor="87b08d">  <td>23/10/12 11:56:29</td>   <td>0</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;BaseClass.setUp()[pri:0, instance:com.LoginTest.VerifyLogin_Nil@4116aac9]">&gt;&gt;setUp</td> 
+<td>&nbsp;</td>  <td>main@989447607</td>   <td></td> </tr>
+<tr bgcolor="87b08d">  <td>23/10/12 11:57:22</td>   <td>52728</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&lt;&lt;BaseClass.tearDown()[pri:0, instance:com.LoginTest.VerifyLogin_Nil@4116aac9]">&lt;&lt;tearDown</td> 
+<td>&nbsp;</td>  <td>main@989447607</td>   <td></td> </tr>
+<tr bgcolor="87b08d">  <td>23/10/12 11:56:29</td>   <td>-759</td> <td title="&gt;&gt;BaseClass.ObjectInit()[pri:0, instance:com.LoginTest.VerifyLogin_Nil@4116aac9]">&gt;&gt;ObjectInit</td> 
+<td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td>  <td>main@989447607</td>   <td></td> </tr>
+<tr bgcolor="a5b572">  <td>23/10/12 11:57:19</td>   <td>49541</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="VerifyLogin_Nil.TC_LoginTest_001_For_Guru99()[pri:0, instance:com.LoginTest.VerifyLogin_Nil@4116aac9]">TC_LoginTest_001_For_Guru99</td> 
+  <td>main@989447607</td>   <td></td> </tr>
+</table>
