+++ test-output/old/MasterTestSuite/reporter-output.html
@@ -1,0 +1,1 @@
+<h2>Reporter output</h2><table></table>
