+++ test-output/old/MasterTestSuite/testng.xml.html
@@ -1,0 +1,1 @@
+<html><head><title>testng.xml for MasterTestSuite</title></head><body><tt>&lt;?xml&nbsp;version="1.0"&nbsp;encoding="UTF-8"?&gt;<br/>&lt;!DOCTYPE&nbsp;suite&nbsp;SYSTEM&nbsp;"https://testng.org/testng-1.0.dtd"&gt;<br/>&lt;suite&nbsp;name="MasterTestSuite"&nbsp;guice-stage="DEVELOPMENT"&gt;<br/>&nbsp;&nbsp;&lt;test&nbsp;thread-count="5"&nbsp;name="MasterTestSuite"&gt;<br/>&nbsp;&nbsp;&nbsp;&nbsp;&lt;classes&gt;<br/>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&lt;class&nbsp;name="com.LoginTest.VerifyLogin_Nil"/&gt;<br/>&nbsp;&nbsp;&nbsp;&nbsp;&lt;/classes&gt;<br/>&nbsp;&nbsp;&lt;/test&gt;&nbsp;&lt;!--&nbsp;MasterTestSuite&nbsp;--&gt;<br/>&lt;/suite&gt;&nbsp;&lt;!--&nbsp;MasterTestSuite&nbsp;--&gt;<br/></tt></body></html>
