+++ test-output/testng-results.xml
@@ -1,73 +1,29 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<testng-results ignored="0" total="1" passed="0" failed="1" skipped="0">
+<testng-results ignored="0" total="1" passed="1" failed="0" skipped="0">
   <reporter-output>
   </reporter-output>
-  <suite started-at="2023-10-11T09:14:54 IST" name="Default suite" finished-at="2023-10-11T09:15:00 IST" duration-ms="6197">
+  <suite started-at="2023-10-12T11:56:29 IST" name="MasterTestSuite" finished-at="2023-10-12T11:57:23 IST" duration-ms="53082">
     <groups>
     </groups>
-    <test started-at="2023-10-11T09:14:54 IST" name="Default test" finished-at="2023-10-11T09:15:00 IST" duration-ms="6197">
-      <class name="com.LoginTest.VerifyLoginTestWithValidCredentials">
-        <test-method is-config="true" signature="ObjectInit()[pri:0, instance:com.LoginTest.VerifyLoginTestWithValidCredentials@7d4f9aae]" started-at="2023-10-11T09:14:53 IST" name="ObjectInit" finished-at="2023-10-11T09:14:54 IST" duration-ms="671" status="PASS">
+    <test started-at="2023-10-12T11:56:29 IST" name="MasterTestSuite" finished-at="2023-10-12T11:57:23 IST" duration-ms="53082">
+      <class name="com.LoginTest.VerifyLogin_Nil">
+        <test-method is-config="true" signature="ObjectInit()[pri:0, instance:com.LoginTest.VerifyLogin_Nil@4116aac9]" started-at="2023-10-12T11:56:29 IST" name="ObjectInit" finished-at="2023-10-12T11:56:29 IST" duration-ms="737" status="PASS">
           <reporter-output>
           </reporter-output>
         </test-method> <!-- ObjectInit -->
-        <test-method is-config="true" signature="setUp()[pri:0, instance:com.LoginTest.VerifyLoginTestWithValidCredentials@7d4f9aae]" started-at="2023-10-11T09:14:54 IST" name="setUp" finished-at="2023-10-11T09:14:57 IST" duration-ms="3560" status="PASS">
+        <test-method is-config="true" signature="setUp()[pri:0, instance:com.LoginTest.VerifyLogin_Nil@4116aac9]" started-at="2023-10-12T11:56:29 IST" name="setUp" finished-at="2023-10-12T11:57:19 IST" duration-ms="49540" status="PASS">
           <reporter-output>
           </reporter-output>
         </test-method> <!-- setUp -->
-        <test-method signature="TC_LoginTest_1_VerifyLoginTestWithValidCrdentials()[pri:0, instance:com.LoginTest.VerifyLoginTestWithValidCredentials@7d4f9aae]" started-at="2023-10-11T09:14:57 IST" name="TC_LoginTest_1_VerifyLoginTestWithValidCrdentials" finished-at="2023-10-11T09:14:59 IST" duration-ms="1833" status="FAIL">
-          <exception class="java.lang.AssertionError">
-            <message>
-              <![CDATA[expected [FaceBook Homepage] but found [Log in to Facebook]]]>
-            </message>
-            <full-stacktrace>
-              <![CDATA[java.lang.AssertionError: expected [FaceBook Homepage] but found [Log in to Facebook]
-at org.testng.Assert.fail(Assert.java:99)
-at org.testng.Assert.failNotEquals(Assert.java:1037)
-at org.testng.Assert.assertEqualsImpl(Assert.java:140)
-at org.testng.Assert.assertEquals(Assert.java:122)
-at org.testng.Assert.assertEquals(Assert.java:629)
-at org.testng.Assert.assertEquals(Assert.java:639)
-at com.LoginTest.VerifyLoginTestWithValidCredentials.TC_LoginTest_1_VerifyLoginTestWithValidCrdentials(VerifyLoginTestWithValidCredentials.java:37)
-at java.base/jdk.internal.reflect.NativeMethodAccessorImpl.invoke0(Native Method)
-at java.base/jdk.internal.reflect.NativeMethodAccessorImpl.invoke(NativeMethodAccessorImpl.java:62)
-at java.base/jdk.internal.reflect.DelegatingMethodAccessorImpl.invoke(DelegatingMethodAccessorImpl.java:43)
-at java.base/java.lang.reflect.Method.invoke(Method.java:566)
-at org.testng.internal.MethodInvocationHelper.invokeMethod(MethodInvocationHelper.java:133)
-at org.testng.internal.TestInvoker.invokeMethod(TestInvoker.java:598)
-at org.testng.internal.TestInvoker.invokeTestMethod(TestInvoker.java:173)
-at org.testng.internal.MethodRunner.runInSequence(MethodRunner.java:46)
-at org.testng.internal.TestInvoker$MethodInvocationAgent.invoke(TestInvoker.java:824)
-at org.testng.internal.TestInvoker.invokeTestMethods(TestInvoker.java:146)
-at org.testng.internal.TestMethodWorker.invokeTestMethods(TestMethodWorker.java:146)
-at org.testng.internal.TestMethodWorker.run(TestMethodWorker.java:128)
-at java.base/java.util.ArrayList.forEach(ArrayList.java:1540)
-at org.testng.TestRunner.privateRun(TestRunner.java:794)
-at org.testng.TestRunner.run(TestRunner.java:596)
-at org.testng.SuiteRunner.runTest(SuiteRunner.java:377)
-at org.testng.SuiteRunner.runSequentially(SuiteRunner.java:371)
-at org.testng.SuiteRunner.privateRun(SuiteRunner.java:332)
-at org.testng.SuiteRunner.run(SuiteRunner.java:276)
-at org.testng.SuiteRunnerWorker.runSuite(SuiteRunnerWorker.java:53)
-at org.testng.SuiteRunnerWorker.run(SuiteRunnerWorker.java:96)
-at org.testng.TestNG.runSuitesSequentially(TestNG.java:1212)
-at org.testng.TestNG.runSuitesLocally(TestNG.java:1134)
-at org.testng.TestNG.runSuites(TestNG.java:1063)
-at org.testng.TestNG.run(TestNG.java:1031)
-at org.testng.remote.AbstractRemoteTestNG.run(AbstractRemoteTestNG.java:115)
-at org.testng.remote.RemoteTestNG.initAndRun(RemoteTestNG.java:251)
-at org.testng.remote.RemoteTestNG.main(RemoteTestNG.java:77)
-]]>
-            </full-stacktrace>
-          </exception> <!-- java.lang.AssertionError -->
+        <test-method signature="TC_LoginTest_001_For_Guru99()[pri:0, instance:com.LoginTest.VerifyLogin_Nil@4116aac9]" started-at="2023-10-12T11:57:19 IST" name="TC_LoginTest_001_For_Guru99" finished-at="2023-10-12T11:57:22 IST" duration-ms="3175" status="PASS">
           <reporter-output>
           </reporter-output>
-        </test-method> <!-- TC_LoginTest_1_VerifyLoginTestWithValidCrdentials -->
-        <test-method is-config="true" signature="tearDown()[pri:0, instance:com.LoginTest.VerifyLoginTestWithValidCredentials@7d4f9aae]" started-at="2023-10-11T09:14:59 IST" name="tearDown" finished-at="2023-10-11T09:15:00 IST" duration-ms="778" status="PASS">
+        </test-method> <!-- TC_LoginTest_001_For_Guru99 -->
+        <test-method is-config="true" signature="tearDown()[pri:0, instance:com.LoginTest.VerifyLogin_Nil@4116aac9]" started-at="2023-10-12T11:57:22 IST" name="tearDown" finished-at="2023-10-12T11:57:23 IST" duration-ms="321" status="PASS">
           <reporter-output>
           </reporter-output>
         </test-method> <!-- tearDown -->
-      </class> <!-- com.LoginTest.VerifyLoginTestWithValidCredentials -->
-    </test> <!-- Default test -->
-  </suite> <!-- Default suite -->
+      </class> <!-- com.LoginTest.VerifyLogin_Nil -->
+    </test> <!-- MasterTestSuite -->
+  </suite> <!-- MasterTestSuite -->
 </testng-results>
